refactor: onboarding page 컴포넌트 분리 완료 (#111)
